// src/app/v2/core/enums/role.enum.ts
export enum Role {
    NATIONAL = 1,
    REGIONAL = 2,
    PROVINCIAL = 3,
    MUNICIPAL = 4,
    PUBLIC = 5,
    DATA_ADMIN = 6,
    REGIONAL_DATA_ADMIN = 7
}

export enum RoleLabel {
    NATIONAL = "National",
    REGIONAL = "Regional",
    PROVINCIAL = "Provincial",
    MUNICIPAL = "Municipal",
    PUBLIC = "Public",
    DATA_ADMIN = "Data Controller",
    REGIONAL_DATA_ADMIN = "Regional Data Controller",
}